判断是否生产环境，方便测试

+++ business/src/main/java/com/study/train/business/controller/ConfirmOrderController.java
@@ -15,6 +15,7 @@ import org.redisson.api.RedissonClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,27 +32,32 @@ public class ConfirmOrderController {
     @Resource
     private BeforeConfirmOrderService beforeConfirmOrderService;
 
+    @Value("${spring.profiles.active}")
+    private String env;
+
     @SentinelResource(value = "confirmOrderDo", blockHandler = "doConfirmBlock")
     @PostMapping("/do")
     public CommonResp<Object> doConfirm(@Valid @RequestBody ConfirmOrderDoReq req) {
 
+        if (!env.equals("dev")) {
+            // 图形验证码校验
+            String imageCodeToken = req.getImageCodeToken();
+            String imageCode = req.getImageCode();
+            RBucket<String> bucket = redissonClient.getBucket(imageCodeToken);
+            String value = bucket.get();
+            LOG.info("从redis中获取到的验证码：{}", value);
+            if (ObjectUtils.isEmpty(value)) {
+                return new CommonResp<>(false, "验证码已过期", null);
+            }
+            // 验证码校验，大小写忽略，提升体验，比如Oo Vv Ww容易混
+            if (!value.equalsIgnoreCase(imageCode)) {
+                return new CommonResp<>(false, "验证码不正确", null);
+            } else {
+                // 验证通过后，移除验证码
+                bucket.delete();
+            }
+        }
 
-        // 图形验证码校验
-        String imageCodeToken = req.getImageCodeToken();
-        String imageCode = req.getImageCode();
-        RBucket<String> bucket = redissonClient.getBucket(imageCodeToken);
-        String value = bucket.get();
-        LOG.info("从redis中获取到的验证码：{}", value);
-        if (ObjectUtils.isEmpty(value)) {
-            return new CommonResp<>(false, "验证码已过期", null);
-        }
-        // 验证码校验，大小写忽略，提升体验，比如Oo Vv Ww容易混
-        if (!value.equalsIgnoreCase(imageCode)) {
-            return new CommonResp<>(false, "验证码不正确", null);
-        } else {
-            // 验证通过后，移除验证码
-            bucket.delete();
-        }
 
         Long id = beforeConfirmOrderService.beforeDoConfirm(req);
         return new CommonResp<>(String.valueOf(id));
